Changes properties of the 'new-item-form' component to line up with the `ember-data-crud` repo.
A `form` property binds directly to the form fields; another property, `newItem`, is a Computed Property which derives its value from `form`, and it is `newItem` that gets passed up through `sendAction` instead of `form`. This intermediation allows for additional logic to be inserted between the form fields and the Route's CRUD action.